README.md
Robust Office Inventory Scan (ROIScan.vbs) is a Microsoft Visual Basic script that collects Office data and provides error detection and analysis options. When you run the script, a log file is created.

The log file can be used to perform tasks such as the following:
•	Assess the installed Office configuration on users' computers.
•	Get detailed data about the Office installation to reproduce the user's environment.
•	Look for mismatched Office versions or configurations.
•	Help troubleshoot Office software updates and installation issues.

After the log file is created, it opens automatically. The default file name is <computer_name>_ROIScan.log and the default location is the current TEMP folder. To navigate to the Temp folder, click Start, click Run, and type %temp%. You must run this command as administrator.

In this article:
•	Download the script
•	Customize and run the script
•	Log file sections
•	FAQs

Download the script:
ROIScan.vbs is available for download in the Office section of the Script Repository in the upload script.

To download the ROIScan.vbs script:
1.	Select Copy Code.
2.	Copy the code to a text editor such as Notepad.
3.	Save the file as ROIScan.vbs.